show code at interpreter start, no special treatment in menu
